Melhorando validações back-end

+++ Cursos/JStack/mycontacts/api/src/app/controllers/ContactController.js
@@ -1,4 +1,6 @@
+const e = require('express');
 const ContactsRepository = require('../repositories/ContactsRepository');
+const isValidUUID = require('../utils/isValidUUID');
 
 class ContactController {
   async index(request, response) {
@@ -12,6 +14,10 @@ class ContactController {
   async show(request, response) {
     // Obter um registro
     const { id } = request.params;
+
+    if(!isValidUUID(id)){
+      return response.status(400).json({error: 'Invalid contact id'})
+    }
 
     const contact = await ContactsRepository.findById(id);
 
@@ -33,13 +39,24 @@ class ContactController {
       return response.status(400).json({ error: 'Name is required' });
     }
 
-    const contactExist = await ContactsRepository.findByEmail(email);
+    if(category_id && !isValidUUID(category_id)){
+      return response.status(400).json({error: 'Invalid category'})
+    }
+
+
+    if (email) {
+      const contactExist = await ContactsRepository.findByEmail(email);
     if (contactExist) {
       response.status(400).json({ error: 'This e-mail is already in use' });
     }
+    }
+
 
     const contact = await ContactsRepository.create({
-      name, email, phone, category_id,
+      name, 
+      email: email || null, 
+      phone, 
+      category_id: category_id || null,
     });
 
     response.status(201).json(contact);
@@ -52,22 +69,35 @@ class ContactController {
       name, email, phone, category_id,
     } = request.body;
 
-    const contactExist = await ContactsRepository.findById(id);
-    if (!contactExist) {
-      return response.status(404).json({ error: 'Contact not found' });
+    if(!isValidUUID(id)){
+      return response.status(400).json({error: 'Invalid contact id'})
+    }
+
+    if(category_id && !isValidUUID(category_id)){
+      return response.status(400).json({error: 'Invalid category'})
     }
 
     if (!name) {
       return response.status(400).json({ error: 'Name is required' });
     }
-
-    const contactByEmail = await ContactsRepository.findByEmail(email);
-    if (contactByEmail && contactByEmail.id !== id) {
-      response.status(400).json({ error: 'This e-mail is already in use' });
+    
+    const contactExist = await ContactsRepository.findById(id);
+    if (!contactExist) {
+      return response.status(404).json({ error: 'Contact not found' });
+    }
+    
+    if (email) {
+      const contactByEmail = await ContactsRepository.findByEmail(email);
+      if (contactByEmail && contactByEmail.id !== id) {
+        response.status(400).json({ error: 'This e-mail is already in use' });
+      }
     }
 
     const contact = await ContactsRepository.update(id, {
-      name, email, phone, category_id,
+      name, 
+      email: email || null, 
+      phone, 
+      category_id: category_id || null,
     });
 
     response.json(contact);
@@ -76,6 +106,10 @@ class ContactController {
   async delete(request, response) {
     // Deletar um registro
     const { id } = request.params;
+
+    if(!isValidUUID(id)){
+      return response.status(400).json({error: 'Invalid contact id'})
+    }
 
     await ContactsRepository.delete(id);
     // 204: No Content
